Code version a2f90b39720df115e21b8f3c6c48db480c1a2c72
Original repository does not play well with composer updates - causes
a lot of dependency conflicts, even if we fix it to this specific
commit. It still fetches the requirements from the latest commit on
the branch. Therefore creating a fork just so that we can pull the
specific commit that we need.